[REFACTOR] 사진 업로드 기능 정리 (#2)
